control-plane/ai/pyproject.toml
[project]
name = "pymes-ai"
version = "0.1.0"
description = "AI assistant service for pymes control-plane"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.27.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.30.0",
  "aiosqlite>=0.20.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sse-starlette>=2.1.0",
  "google-genai>=1.0.0",
  "python-jose[cryptography]>=3.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
